pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "scalar_email_triage"
version = "0.1.0"
description = "Email Triage RL Environment for Meta OpenEnv"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "numpy",
    "openai",
    "openenv-core>=0.2.0"
]

[project.scripts]
server = "server.app:main"

